Rework of the DTP module to fit into VMEM requests (#1)
* Initial work done on re-factoring the DTP server.
Introduced a asynchronous model for the DTP server.

* Added DTP VMEM client - WIP.

Introduced V2 of interval_t and dtp_meta_req_t types to be able to handle DTP through VMEM.

* Changed the meta member of the dtp_msg_t type to be v2 in order to accept 64-bit addresses.
Limited the number of intervals in one dtp meta request (v2) to be 8.

* Added base address to the meta data.
Adapted the handling of the intervals.
Cleaned up the code a bit.
Reverted the dtp V2 changes, and added vaddr and size to the dtp_vmem_request_t type.

* Added external declaration of vmem_dtp_download() method.

* Encapsulated the DTP requests into a deeper level of nesting into the VMEM request. This will result in we only use/occupy one VMEM user request (0x80).
Enclosed the entire DTP receiver into the DTP VMEM client.

* Added the endianness control to the DTP requests.
Implemented the stop download feature.

* Added a VMEM DTP server handling mechanism.

* Changed to use libossi rather then libsi for access to message_queue

* Using libossi rather than libsi.

* Moved out the target specific handling of the DTP on_data handling. This is now located in the csh APM.

* Removed link_as_whole() from dependency_declare, and partial dependencies from csp, param and ossi.